+++ compass/src/config/luigi-config/navitagion.js
@@ -1,15 +1,4 @@
-const getTenants = () => {
-  const tenants = [{
-    category: 'Tenants',
-    label: 'Tenant 1',
-    pathValue: 'tenant1'
-  }, {
-    category: 'Tenants',
-    label: 'Tenant 2',
-    pathValue: 'tenant2'
-  }];
-  return tenants
-}
+import { getTenants } from './helpers/navigation-helpers'
 
 const navigation = {
   nodes: () => [
@@ -19,37 +8,30 @@ const navigation = {
       label: 'Overview',
       viewUrl: 'http://localhost:3000/',
     },
-
     {
       hideSideNav: true,
       hideFromNav: true,
       pathSegment: 'tenant',
-      label: 'Tenants',
       children: [
         {
           hideSideNav: true,
           pathSegment: ':tenantId',
-          label: 'Tenant',
           navigationContext: 'tenant',
           context: {
             tenantId: ':tenantId'
           },
           children: [
             {
-              hideSideNav: true,
+              keepSelectedForChildren: true,
               pathSegment: 'runtimes',
               label: 'Runtimes',
               viewUrl: 'http://localhost:3000/runtimes',
             },
             
             {
-              hideSideNav: true,
-              hideFromNav: true,
               pathSegment: 'runtime',
               children: [
                 {
-                  hideSideNav: true,
-                  hideFromNav: true,
                   pathSegment: ':runtimeId',
                   label: 'Runtimes',
                   viewUrl: 'http://localhost:3000/runtime/:runtimeId',
@@ -65,7 +47,7 @@ const navigation = {
   contextSwitcher: {
     defaultLabel: 'Select Tenant...',
     parentNodePath: '/tenant', 
-    lazyloadOptions: true, // load options on click instead on page load
+    lazyloadOptions: true,
     options: getTenants,
   }
 };
